--- Assets/Scripts/Runtime/CH2/IntroDialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 233ac0396113eb54c9ae345a5d3d8d26
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
