+++ Assets/Scripts/Runtime/CH2/Main.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f335063f422aa28488098c46ac5054d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
