+++ Assets/Scripts/Runtime/CH2/Main/CH2Controller.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+using Sound = Runtime.ETC.Sound;
+
+namespace Runtime.CH2
+{
+    public class CH2Controller : MonoBehaviour
+    {
+        void Start()
+        {
+            Managers.Sound.Play(Sound.BGM, "Mamago_BGM_1");
+        }
+    }
+}
