+++ Assets/Scripts/Runtime/CH2/Main/CH2Controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9fca31c22b06ef74fa2204c4dde43a85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
