+++ Assets/Scripts/Runtime/CH2/Main/CH2KeyBinder.cs
@@ -1,0 +1,28 @@
+using Runtime.Common.View;
+using UnityEngine;
+
+namespace Runtime.CH2
+{
+    public class CH2KeyBinder : MonoBehaviour
+    {
+        [SerializeField]
+        private SettingsUIView settingsUIView;
+
+        private void Start()
+        {
+            InitKeyBinding();
+        }
+
+        private void InitKeyBinding()
+        {
+            Managers.Data.InGameKeyBinder.GameControlReset();
+            Managers.Data.InGameKeyBinder.CH2KeyBinding(this);
+        }
+
+        public void SetSettingUI()
+        {
+            settingsUIView.GameSettingToggle();
+            Time.timeScale = (Time.timeScale == 0 ? 1 : 0);
+        }
+    }
+}
