+++ Assets/Scripts/Runtime/CH2/Main/CH2KeyBinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c74774b2cbc103c4694e3a9834df7e29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
