+++ Assets/Scripts/Runtime/Input/InGameKeyBinder.cs
@@ -1,6 +1,7 @@
 using Runtime.CH1.Main.Player;
 using Runtime.CH1.Pacmom;
 using Runtime.CH1.Title;
+using Runtime.CH2;
 using Runtime.Common.View;
 
 namespace Runtime.Input
@@ -52,7 +53,14 @@ namespace Runtime.Input
             _gameOverControls.UI.Enable();
             _gameOverControls.UI.GameSetting.performed += _ => settingsUIView.GameSettingToggle();
         }
-        
+
+        // CH2
+        public void CH2KeyBinding(CH2KeyBinder keyBinder)
+        {
+            _gameOverControls.UI.Enable();
+            _gameOverControls.UI.GameSetting.performed += _ => keyBinder.SetSettingUI();
+        }
+
         // ETC
         public void GameControlReset()
         {
